Added JwtAuthenticationFilter and AuthenticationController.

+++ src/main/resources/db/migration/V1__.sql
@@ -122,7 +122,7 @@ CREATE TABLE IF NOT EXISTS users
     username             VARCHAR(255)                            NOT NULL,
     email                VARCHAR(255)                            NOT NULL,
     password             VARCHAR(255)                            NOT NULL,
-    verification_code    INTEGER,
+    verification_code    BIGINT,
     verification_expires time WITHOUT TIME ZONE,
     enabled              BOOLEAN,
     CONSTRAINT users_pkey PRIMARY KEY (id)
